+++ extension/src/panel/pages/routes/edit.$routeId/AvatarInput.tsx
@@ -1,15 +1,7 @@
-import {
-  Blockie,
-  GhostButton,
-  Input,
-  selectStyles,
-  TextInput,
-} from '@/components'
+import { Blockie, Input, Select, selectStyles, TextInput } from '@/components'
 import { validateAddress } from '@/utils'
 import { getAddress } from 'ethers'
-import { ChevronDown, X } from 'lucide-react'
 import { useEffect, useState } from 'react'
-import type { CommonProps, GroupBase } from 'react-select'
 import CreatableSelect from 'react-select/creatable'
 import type { Option } from './ModSelect'
 
@@ -42,7 +34,10 @@ export const AvatarInput = ({
               inputId={inputId}
               blurInputOnSelect
               isClearable
-              components={{ ClearIndicator, DropdownIndicator }}
+              components={{
+                ClearIndicator: Select.ClearIndicator<Option, false>,
+                DropdownIndicator: Select.DropdownIndicator,
+              }}
               formatOptionLabel={SafeOptionLabel}
               placeholder="Paste an address or select from the list"
               classNames={selectStyles<{ value: string; label: string }>()}
@@ -106,17 +101,3 @@ const SafeOptionLabel = (option: Option) => {
     </div>
   )
 }
-
-const ClearIndicator = ({
-  clearValue,
-}: CommonProps<Option, false, GroupBase<Option>>) => (
-  <GhostButton iconOnly icon={X} size="small" onClick={clearValue}>
-    Clear piloted Safe
-  </GhostButton>
-)
-
-const DropdownIndicator = () => (
-  <GhostButton iconOnly icon={ChevronDown} size="small">
-    View available Safes
-  </GhostButton>
-)
